+++ orchestrator/orchestrator.go
@@ -29,6 +29,7 @@ import (
 	"github.com/luowensheng/perch/usecases/installlsp"
 	"github.com/luowensheng/perch/usecases/installvscode"
 	"github.com/luowensheng/perch/usecases/importsh"
+	"github.com/luowensheng/perch/usecases/scan"
 	"github.com/luowensheng/perch/usecases/listcommands"
 	"github.com/luowensheng/perch/usecases/runbuild"
 	"github.com/luowensheng/perch/usecases/runcommand"
@@ -437,6 +438,7 @@ func buildCLI(r ops.Restrictions, envAllow, allowBins map[string]bool, noMeta bo
 		InstallLSP:    &installlsp.Impl{},
 		InstallVSCode: &installvscode.Impl{InstallLSP: (&installlsp.Impl{}).Execute},
 		ImportSh:      &importsh.Impl{},
+		Scan:          &scan.Impl{Load: capyloader.Load},
 	}
 
 	return &cli.CLI{
@@ -516,6 +518,7 @@ func buildEmbeddedCLI(bundle *embed.Bundle, r ops.Restrictions, envAllow, allowB
 		InstallLSP:    &installlsp.Impl{},
 		InstallVSCode: &installvscode.Impl{InstallLSP: (&installlsp.Impl{}).Execute},
 		ImportSh:      &importsh.Impl{},
+		Scan:          &scan.Impl{Load: capyloader.Load},
 	}
 
 	version := p.Version